pyproject.toml
[project]
name = "chatbot-demo"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "chromadb>=1.5.5",
    "openai>=2.28.0",
    "pymupdf>=1.27.2",
    "python-dotenv>=1.2.2",
    "tiktoken>=0.12.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "ruff>=0.15.6",
]

[tool.pytest.ini_options]
testpaths = [
    "tests"
]
pythonpath = [
    "."
]

[tool.ruff]
line-length = 120
